BigReal er klargjort til x64. Manngler at skrive shortProduct som bruger 64-bit-registrene, samt test. Virker stadig som før i x32

+++ Lib/Src/BigReal/BigInt.cpp
@@ -93,7 +93,7 @@ BigInt &BigInt::operator/=(const BigInt &x) {
   return *this = *this / x;
 }
 
-BigInt randomInteger(int length, Random *rnd, DigitPool *pool) {
+BigInt randomInteger(size_t length, Random *rnd, DigitPool *pool) {
   return BigInt(e(BigReal::random(length, rnd, pool),length,pool),pool);
 }
